back-end/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
alembic
bcrypt
python-dotenv